Move NotificationScope to standalone file
